Make num_bufs a command line argument (plus related fixes)

+++ lib/include/quant/quant.h
@@ -47,7 +47,7 @@ struct q_conf {
     const char * const tls_cert;     // required for server
     const char * const tls_key;      // required for server
     const char * const tls_log;
-    const uint32_t num_bufs;
+    const uint64_t num_bufs;
     const uint8_t enable_tls_cert_verify : 1;
     uint8_t : 7;
 };
